# openapi/schema.yml
openapi: 3.1.0
info:
  title: Dockyard API
  version: 1.0.0 (v1)
paths:
...
  /api/projects/:
    get:
      summary: List Projects
      operationId: list_projects_api_projects__get
      parameters:
      - name: slug
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Slug
      - name: sort_by
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Sort By
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProjectSchema'
                title: Response List Projects Api Projects  Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema: *id001
        default:
          description: Error
          content:
            application/json:
              schema: *id001
    post:
      summary: Create Project
      operationId: create_project_api_projects__post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProjectCreateRequest'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectSchema'
        '422':
          description: Validation Error
          content:
            application/json:
              schema: *id001
        default:
          description: Error
          content:
            application/json:
              schema: *id001
  /api/projects/{slug}/:
    get:
      summary: Get Project
      operationId: get_project_api_projects__slug___get
      parameters:
      - name: slug
        in: path
        required: true
        schema:
          type: string
          title: Slug
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectSchema'
        '422':
          description: Validation Error
          content:
            application/json:
              schema: *id001
        default:
          description: Error
          content:
            application/json:
              schema: *id001
    put:
      summary: Update Project
      operationId: update_project_api_projects__slug___put
      parameters:
      - name: slug
        in: path
        required: true
        schema:
          type: string
          title: Slug
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProjectUpdateRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectSchema'
        '422':
          description: Validation Error
          content:
            application/json:
              schema: *id001
        default:
          description: Error
          content:
            application/json:
              schema: *id001
    delete:
      summary: Archive Project
      operationId: archive_project_api_projects__slug___delete
      parameters:
      - name: slug
        in: path
        required: true
        schema:
          type: string
          title: Slug
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema: *id001
        default:
          description: Error
          content:
            application/json:
              schema: *id001
  /api/projects/{project_slug}/environments/:
    post:
      summary: Create Environment
      operationId: create_environment_api_projects__project_slug__environments__post
      parameters:
      - name: project_slug
        in: path
        required: true
        schema:
          type: string
          title: Project Slug
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateEnvironmentRequest'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SimpleEnvironmentSchema'
        '422':
          description: Validation Error
          content:
            application/json:
              schema: *id001
        default:
          description: Error
          content:
            application/json:
              schema: *id001
  /api/projects/{project_slug}/environments/{env_slug}/:
    delete:
      summary: Delete Environment
      operationId: delete_environment_api_projects__project_slug__environments__env_slug___delete
      parameters:
      - name: project_slug
        in: path
        required: true
        schema:
          type: string
          title: Project Slug
      - name: env_slug
        in: path
        required: true
        schema:
          type: string
          title: Env Slug
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema: *id001
        default:
          description: Error
          content:
            application/json:
              schema: *id001
  /api/projects/{project_slug}/environments/{env_slug}/variables/:
    get:
      summary: List Environment Variables
      operationId: list_environment_variables_api_projects__project_slug__environments__env_slug__variables__get
      parameters:
      - name: project_slug
        in: path
        required: true
        schema:
          type: string
          title: Project Slug
      - name: env_slug
        in: path
        required: true
        schema:
          type: string
          title: Env Slug
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SharedEnvVariableSchema'
                title: Response List Environment Variables Api Projects  Project Slug  Environments  Env
                  Slug  Variables  Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema: *id001
        default:
          description: Error
          content:
            application/json:
              schema: *id001
    post:
      summary: Create Environment Variable
      operationId: create_environment_variable_api_projects__project_slug__environments__env_slug__variables__post
      parameters:
      - name: project_slug
        in: path
        required: true
        schema:
          type: string
          title: Project Slug
      - name: env_slug
        in: path
        required: true
        schema:
          type: string
          title: Env Slug
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SharedEnvVariableRequest'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SharedEnvVariableSchema'
        '422':
          description: Validation Error
          content:
            application/json:
              schema: *id001
        default:
          description: Error
          content:
            application/json:
              schema: *id001
  /api/projects/{project_slug}/environments/{env_slug}/variables/{variable_id}/:
    delete:
      summary: Delete Environment Variable
      operationId: delete_environment_variable_api_projects__project_slug__environments__env_slug__variables__variable_id___delete
      parameters:
      - name: project_slug
        in: path
        required: true
        schema:
          type: string
          title: Project Slug
      - name: env_slug
        in: path
        required: true
        schema:
          type: string
          title: Env Slug
      - name: variable_id
        in: path
        required: true
        schema:
          type: string
          title: Variable Id
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema: *id001
        default:
          description: Error
          content:
            application/json:
              schema: *id001
...
components:
  schemas:
...
    CreateEnvironmentRequest:
      properties:
        name:
          type: string
          maxLength: 255
          minLength: 1
          pattern: ^[a-zA-Z0-9]+(?:-[a-zA-Z0-9]+)*$
          title: Name
      type: object
      required:
      - name
      title: CreateEnvironmentRequest
...
    ProjectCreateRequest:
      properties:
        slug:
          anyOf:
          - type: string
            maxLength: 255
            pattern: ^[-a-zA-Z0-9_]+$
          - type: 'null'
          title: Slug
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
      type: object
      title: ProjectCreateRequest
    ProjectSchema:
      properties:
        id:
          type: string
          title: Id
        slug:
          type: string
          title: Slug
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
        created_at:
          type: string
          format: date-time
          title: Created At
        updated_at:
          type: string
          format: date-time
          title: Updated At
        environments:
          items:
            $ref: '#/components/schemas/SimpleEnvironmentSchema'
          type: array
          title: Environments
        healthy_services:
          type: integer
          title: Healthy Services
          default: 0
        total_services:
          type: integer
          title: Total Services
          default: 0
        healthy_stack_services:
          type: integer
          title: Healthy Stack Services
          default: 0
        total_stack_services:
          type: integer
          title: Total Stack Services
          default: 0
      type: object
      required:
      - id
      - slug
      - description
      - created_at
      - updated_at
      - environments
      title: ProjectSchema
    ProjectUpdateRequest:
      properties:
        slug:
          anyOf:
          - type: string
            maxLength: 255
            pattern: ^[-a-zA-Z0-9_]+$
          - type: 'null'
          title: Slug
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
      type: object
      title: ProjectUpdateRequest
...
    SharedEnvVariableRequest:
      properties:
        key:
          type: string
          maxLength: 255
          minLength: 1
          title: Key
        value:
          type: string
          title: Value
          default: ''
      type: object
      required:
      - key
      title: SharedEnvVariableRequest
    SharedEnvVariableSchema:
      properties:
        id:
          type: string
          title: Id
        key:
          type: string
          title: Key
        value:
          type: string
          title: Value
      type: object
      required:
      - id
      - key
      - value
      title: SharedEnvVariableSchema
    SimpleEnvironmentSchema:
      properties:
        id:
          type: string
          title: Id
        name:
          type: string
          title: Name
        is_preview:
          type: boolean
          title: Is Preview
        created_at:
          type: string
          format: date-time
          title: Created At
      type: object
      required:
      - id
      - name
      - is_preview
      - created_at
      title: SimpleEnvironmentSchema
...